refactor: move SequenceSplitter to sub_document.py
